stt_worker/stt_worker_base/requirements.txt
prometheus-client
kafka-python
minio
torch==2.1.0
torchaudio==2.1.0
openai-whisper
numpy<2.0
tqdm
tiktoken
pydub